Move `impl StakingQuery for BaseCoinApp` from app/mod.rs -> app/modules/staking.rs (#38)
